requirements.txt
kafka-python>=2.0.2
pyspark>=3.1.2
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=0.24.2